refactor: 강의평가 작성 컴포넌트 2차 리팩토링
